api/requirements.txt
Pillow==9.5.0
numpy==1.24.3
onnxruntime==1.16.0
